fix(calc): Fix bug stopping calculator from evaluating equations
Bug:
When called, calculate() checks if leftNum, rightNum, and
operator are set. If they are, then they are evaluated and leftNum is
set to the result.

However, this result is a number, meaning it can be 0. When this happens
leftNum always evaluates to false. Since leftNum cannot be changed once
an operator is set, this prevents the user from evaluating any other
calculations.

Fix:
To fix the issue, calculation results are converted to
strings before they are set as leftNum. This allows calculations with 0
as a result to be evaluated as "0", which returns true.