fix(calc): Fix bug preventing user from entering leftNum
Bug:
handleInputDigit only updates leftNum if no operator is set. Once an
operator is set, leftNum cannot be updated. If an operator is set before
leftNum has a value, it will remain an empty string and prevent any
calculations from being performed.

Fix:
A check was added to ensure there is a value for leftNum before
preventing it from being updated.