refac(calc): Rework global calculator variables for refactor
Rename variables to be more accurate to upcoming rework, where the
calculator will use focusNum and storedNum instead of leftNum and
rightNum. This change will allow most functions to work exclusively with
focusNum, eliminating the need to check whether leftNum or rightNum is
being modified and massively simplifying the functions.

This rework is based on Zonwrk's calculator:
https://github.com/Zonwrk/odin-calculator